test: 테스틐 코드 클래스명 변경 및 불필요한 테스트 메서드 삭제
